🔥 feat: Add support for embedded Koa-Style Req and Res structs (#3533)
